requirements.txt
numpy
matplotlib
pillow
tensorflow
keras
scikit-learn
mtcnn
opencv-python
keras-facenet